code/requirements.txt
torch
torchvision
compressai>=1.2
pytorch-msssim
Pillow
numpy
tensorboard
git+https://github.com/openai/CLIP.git
